+++ packages/ko-lint-config/index.js
@@ -2,9 +2,8 @@ const prettier = require('prettier');
 const eslint = require('eslint');
 const stylelint = require('stylelint');
 
-
 module.exports = {
   prettier,
   eslint,
-  stylelint
-}
+  stylelint,
+};