+++ packages/ko-lint-config/test/basic.js
@@ -3,13 +3,13 @@ const config = require('../');
 const test = require('tape');
 
 test('test basic properties of config', function (t) {
-    t.ok(isObject(config.parserOptions));
-    t.ok(isObject(config.env));
-    t.ok(isObject(config.globals));
-    t.ok(isObject(config.rules));
-    t.end();
+  t.ok(isObject(config.parserOptions));
+  t.ok(isObject(config.env));
+  t.ok(isObject(config.globals));
+  t.ok(isObject(config.rules));
+  t.end();
 });
 
 function isObject(obj) {
-    return typeof obj === 'object' && obj !== null;
+  return typeof obj === 'object' && obj !== null;
 }