+++ packages/ko-lints/index.d.ts
@@ -4,6 +4,8 @@ export type IOpts = {
   write: boolean;
   configPath: string;
   patterns: Pattern[];
+  concurrency?: boolean;
+  concurrentNumber?: number;
 };
 
 export type IKeys = 'eslint' | 'prettier' | 'stylelint';
@@ -13,7 +15,7 @@ declare class Lints {
   private opts;
   state: 0 | 1;
   constructor(opts: IOpts);
-  run(key: IKeys): Promise<true | string[]>;
+  run(key: IKeys): Promise<string[]>;
 }
 
 export default Lints;