+++ packages/ko-lints/src/interfaces.ts
@@ -1,9 +1,27 @@
 import { Pattern } from 'fast-glob';
+import Parser from './factory/parser';
 
 export type IOpts = {
   write: boolean;
   configPath: string;
   patterns: Pattern[];
+  concurrency?: boolean;
+  concurrentNumber?: number;
 };
 
 export type IKeys = 'eslint' | 'prettier' | 'stylelint';
+
+export type IParserOpts = Pick<IOpts, 'write' | 'configPath'> & {
+  name: IKeys;
+};
+
+export type IThreadOpts = IParserOpts & {
+  entries: string[];
+  concurrentNumber: NonNullable<IOpts['concurrentNumber']>;
+};
+
+export type IChildOpts = {
+  parser: Parser;
+  ignoreFiles: string[];
+  name: IKeys;
+};