+++ packages/ko/CHANGELOG.md
@@ -1,5 +1,49 @@
 # ko
 
+## 6.1.2
+
+### Patch Changes
+
+- ff1560a2: exit with error when lint throw errors
+
+  - exit with `process.exit(1)` when stdout length is not zero
+  - use variadic arguments in patterns
+
+## 6.1.1
+
+### Patch Changes
+
+- 62572f85: update dependence of ko-lints
+
+## 6.1.0
+
+### Minor Changes
+
+- 1721dfa0: Support Lints & Format with concurrency mode, Add Config & Plugin of Prettier in Eslint
+
+  - `ko-lint-config`:
+
+    Add Config & Plugin of Prettier in Eslint Config
+
+  - `ko-lints`:
+
+    - Support concurrency mode with multithreading to do lint or format tasks,you can enable concurrency mode via `--concurrency` cli flag
+    - Default threads Count is `require('os').cpus().length`, you can specify it via `--concurrentNumber` like `--concurrentNumber=4`
+
+  - `ko`:
+    - Support `ko prettier`,`ko eslint`,`ko stylelint` commands with concurrency mode, you can also specify this mode with `ko.config.js`
+
+### Patch Changes
+
+- Updated dependencies [1721dfa0]
+  - ko-lints@4.0.0
+
+## 6.0.1
+
+### Patch Changes
+
+- ko-lints@3.0.0
+
 ## 6.0.0
 
 ### Major Changes